rename examples to test_examples (#255)
* rename examples to test_examples

* rename

* fix tests

* fix tests

* fix tests
