+++ src/litserve/__init__.py
@@ -11,10 +11,10 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from litserve.__about__ import *  # noqa: F401, F403
+from litserve.__about__ import *  # noqa: F403
 from litserve.api import LitAPI
 from litserve.server import LitServer, Request, Response
-from litserve import examples
+from litserve import test_examples
 from litserve.specs.openai import OpenAISpec
 
-__all__ = ["LitAPI", "LitServer", "Request", "Response", "examples", "OpenAISpec"]
+__all__ = ["LitAPI", "LitServer", "Request", "Response", "test_examples", "OpenAISpec"]
