+++ tests/e2e/default_api.py
@@ -14,6 +14,6 @@
 import litserve as ls
 
 if __name__ == "__main__":
-    api = ls.examples.SimpleLitAPI()
+    api = ls.test_examples.SimpleLitAPI()
     server = ls.LitServer(api)
     server.run(port=8000)
