+++ tests/e2e/default_batching.py
@@ -14,6 +14,6 @@
 import litserve as ls
 
 if __name__ == "__main__":
-    api = ls.examples.SimpleBatchedAPI()
+    api = ls.test_examples.SimpleBatchedAPI()
     server = ls.LitServer(api, max_batch_size=4, batch_timeout=0.05)
     server.run(port=8000)
